:hammer: Move home intent
